Fold GCS client setup into storage
